server/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
h11==0.14.0
idna==3.4
plumbum==1.8.1
pydantic==1.10.5
python-dotenv==1.0.0
python-multipart==0.0.5
rpyc==5.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
